Add non-babel render benchmarks
