+++ .gitignore
@@ -13,6 +13,7 @@ src/React.Sample.Webpack/build
 src/React.Sample.CoreMvc/wwwroot/js/Sample.js
 *.lock.json
 .vs/
+results/
 
 ## Ignore Visual Studio temporary files, build results, and
 ## files generated by popular Visual Studio add-ons.
