+++ tests/React.Tests.Benchmarks/React.Tests.Benchmarks.csproj
@@ -29,6 +29,9 @@
     <None Update="Sample.jsx">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
+	<None Update="Sample.js">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 
 </Project>
