+++ tests/React.Tests.Benchmarks/Sample.js
@@ -1,0 +1,9 @@
+function HelloWorld(props) {
+	return React.createElement(
+		"div",
+		null,
+		"Hello ",
+		props.name,
+		"!"
+	);
+}
